Move to slogadapter package
