+++ internal/component/otelcol/auth/auth.go
@@ -12,19 +12,20 @@ import (
 	"hash/fnv"
 	"strings"
 
-	"github.com/grafana/alloy/internal/component"
-	otelcolCfg "github.com/grafana/alloy/internal/component/otelcol/config"
-	"github.com/grafana/alloy/internal/component/otelcol/internal/lazycollector"
-	"github.com/grafana/alloy/internal/component/otelcol/internal/scheduler"
-	otelcolutil "github.com/grafana/alloy/internal/component/otelcol/util"
-	"github.com/grafana/alloy/internal/util/zapadapter"
-	"github.com/grafana/alloy/syntax"
 	"github.com/prometheus/client_golang/prometheus"
 	otelcomponent "go.opentelemetry.io/collector/component"
 	otelextension "go.opentelemetry.io/collector/extension"
 	"go.opentelemetry.io/collector/pipeline"
 	sdkprometheus "go.opentelemetry.io/otel/exporters/prometheus"
 	"go.opentelemetry.io/otel/sdk/metric"
+
+	"github.com/grafana/alloy/internal/component"
+	otelcolCfg "github.com/grafana/alloy/internal/component/otelcol/config"
+	"github.com/grafana/alloy/internal/component/otelcol/internal/lazycollector"
+	"github.com/grafana/alloy/internal/component/otelcol/internal/scheduler"
+	otelcolutil "github.com/grafana/alloy/internal/component/otelcol/util"
+	"github.com/grafana/alloy/internal/slogadapter"
+	"github.com/grafana/alloy/syntax"
 )
 
 var (
@@ -231,7 +232,7 @@ func (a *Auth) Update(args component.Arguments) error {
 	settings := otelextension.Settings{
 		ID: otelcomponent.NewIDWithName(a.factory.Type(), a.opts.ID),
 		TelemetrySettings: otelcomponent.TelemetrySettings{
-			Logger:         zapadapter.New(a.opts.Logger),
+			Logger:         slogadapter.New(a.opts.Logger),
 			TracerProvider: a.opts.Tracer,
 			MeterProvider:  mp,
 		},
